+++ src/assets/styles/dashboard.css
@@ -7,10 +7,10 @@
 .chart-wrap{ margin-top: 40px; padding: 20px 40px; display: flex; justify-content: space-between; align-items: center; flex-wrap: wrap; background-color: #f1f1f1;}
 .chart-wrap .date-box{ border-top:1px solid #000;}
 .chart-box{width: 48%; position: relative; margin-bottom: 60px; height: 380px; background-color: #fff;}
-<<<<<<< HEAD
+
 .chart-wrap canvas{ margin: 0 auto;}
 .custom-legend{display: flex; align-items: center; justify-content: center}
-=======
+
 .chart-box.wide{width: 100%; height: auto;}
 .chart-box.wide canvas{margin:0; padding: 0 20px;}
 .chart-box .title{padding:10px 20px; display: flex; align-items: center; justify-content: space-between;}
@@ -18,7 +18,6 @@
 .custom-legend{display: flex; align-items: center; justify-content: center}
 .custom-legend.vertical { width: 100px; flex-wrap: wrap; justify-content: flex-start; position: absolute; right: 50px; top: 80px;}
 .custom-legend.vertical .legend-item{margin-right: 0; margin-bottom: 2px;}
->>>>>>> 2c92b41 (yus)
 .custom-legend .legend-item{display: flex; align-items: center; margin-right: 10px;}
 .custom-legend .legend-item .color-box{width: 10px; height: 10px; margin-right: 5px; padding:0;}
 .custom-legend .legend-item .label-item{margin-right: 15px;}
